+++ src/tedi/components/buttons/button-content/button-content.tsx
@@ -144,14 +144,12 @@ const InternalButtonContent = forwardRef(
     const getIcon = (location: string, icon: string | IconWithoutBackgroundProps): JSX.Element => {
       const iconBEM = cn(styles['tedi-btn__icon'], styles[`tedi-btn__icon--${location}`], {
         [styles['tedi-btn__spinner']]: isLoading,
+        [styles['tedi-btn__icon-responsive']]: Component === 'a' && size === 'small',
       });
-
-      const isLink = visualType === 'link';
 
       const defaultIconProps: Partial<IconWithoutBackgroundProps> = {
         size: size === 'large' ? 24 : 18,
         className: iconBEM,
-        ...(isLink ? { display: 'inline' } : {}),
       };
 
       const iconProps: IconWithoutBackgroundProps =